Update admin_theme.php
The only way to see OK the menus from frontend of the theme in Admin Area/Menu Manager is to load first bootstrap 3.3.5 and after that old bootstrap 2 for admin area. I changed also the order of files loading, first css files and then js files. All seams to work OK on frontend with bootstrap 3.3.5 and in admin area with old bootstrap 2